refactor(common): add commend and move packages
